+++ src/reader-activation-newsletters/newsletters-modal.js
@@ -1,3 +1,5 @@
+import * as a11y from '../reader-activation-auth/accessibility.js';
+
 /**
  * Get the newsletters signup modal container.
  *
@@ -63,7 +65,9 @@ export function openNewslettersSignupModal( config = {} ) {
 	container.config = config;
 
 	container.newslettersSignupCallback = ( message, data ) => {
-		close();
+		if ( config?.closeOnSuccess ) {
+			close();
+		}
 		if ( config?.callback ) {
 			config.callback( message, data );
 		}
@@ -91,6 +95,7 @@ export function openNewslettersSignupModal( config = {} ) {
 	}
 
 	modal.setAttribute( 'data-state', 'open' );
+	a11y.trapFocus( modal );
 	if ( window?.newspackReaderActivation?.overlays ) {
 		modal.overlayId = window.newspackReaderActivation.overlays.add();
 	}